+++ context/app/static/js/App.jsx
@@ -25,6 +25,7 @@ function App(props) {
 
 App.propTypes = {
   flaskData: PropTypes.exact({
+    title: PropTypes.string,
     entity: PropTypes.object,
     flashed_messages: PropTypes.array,
     provenance: PropTypes.object,